long long Time;

+++ p2psim/event.C
@@ -21,7 +21,7 @@ Event::~Event()
 Event::Event(vector<string> *v)
 {
   _id = _uniqueid++;
-  this->ts = (Time) atoi((*v)[0].c_str());
+  this->ts = (Time) atoll((*v)[0].c_str());
   v->erase(v->begin());
 }
